+++ packages/widgetbook_annotation/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- **FEAT**: Add `AddonConfig` for [Semantics Addon](https://docs.widgetbook.io/addons/semantics-addon). ([#1402](https://github.com/widgetbook/widgetbook/pull/1402))
+
 ## 3.3.1
 
 - **REFACTOR**: Add generic parameter to `AddonConfig`; to allow supporting non-primitive addon configs (e.g. `ViewportAddonConfig`). ([#1318](https://github.com/widgetbook/widgetbook/pull/1318))